added dates to add, list, moved list func, use f64 for schema version

+++ src/handler/data.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize, Serialize, Clone)]
 pub struct DirectoryInfo {
-    pub version: u64,
+    pub version: f64,
     pub name: String,
 }
